Adding some size_t correctness for GCC, also fixing a time bug. The otim/time variables would be uninitialized if the XBoard interface never gave a time, causing our time function to possibly return a negative time allocation. In turn, this caused the iterative deepening function to bail out excessively early, leading to uninitialized moves in a move buffer that cause segfaults when printing them as notation.

+++ src/book.c
@@ -77,7 +77,7 @@ static int readMoves(GameState* state, char*** strArray, OpenBook* book) {
 	FILE *fp;
 	fp = popen(sql, "r");
 	if (fp) {
-		while (fgets(line, lineLen, fp) && count < BOOK_MAX_MOVES) {
+		while (fgets(line, (int) lineLen, fp) && count < BOOK_MAX_MOVES) {
 			char* idx = line + strlen(line) - 1;
 			while (idx > line && iscntrl(*idx)) {
 				*idx-- = '\0';
